packages/pixel-patrol-base/pyproject.toml
[project]
name = "pixel-patrol-base"
version = "0.1.0"
description = "Image prevalidation tool"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.1",
    "dash-ag-grid>=32.3.0",
    "dash-bootstrap-components>=2.0.3",
    "matplotlib>=3.10.3",
    "polars>=1.31.0",
    "pyarrow>=21.0.0",
    "pyyaml>=6.0.2",
    "statsmodels>=0.14.4",
    "tensorboard>=2.18.0",
    "tensorboardx>=2.6.4",
    "zarr>=3.1.0",
]

[tool.uv]
package = true


[tool.hatch.build.targets.wheel]
packages = ["src/pixel_patrol_base"]

[project.scripts]
pixel-patrol-base = "pixel_patrol_base.cli:cli"

[project.entry-points."pixel_patrol.widget_plugins"]
pixel_patrol_base_widget_builtins = "pixel_patrol_base.plugin_registry:register_widget_plugins"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-mock>=3.14.1",
]
